server/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
httpx==0.26.0
python-dotenv==1.0.0
openai==1.12.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.0
bcrypt==4.0.1
